+++ crates/pg_base_db/src/path.rs
@@ -20,3 +20,4 @@ impl PgLspPath {
         }
     }
 }
+
