+++ crates/pg_lsp/src/client/client_config_opts.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+// TODO: Check that the Opts are correct (existed in server.rs)
+#[derive(Deserialize)]
+pub struct ClientConfigurationOptions {
+    pub db_connection_string: Option<String>
+}
