+++ crates/pg_lsp/src/session.rs
@@ -1,20 +1,22 @@
-use std::{collections::HashSet, sync::Arc};
+use std::{collections::HashSet,sync::Arc};
 
 use pg_base_db::{Change, DocumentChange, PgLspPath};
-use pg_commands::ExecuteStatementCommand;
+use pg_commands::{Command, ExecuteStatementCommand};
+use pg_completions::CompletionParams;
 use pg_diagnostics::Diagnostic;
 use pg_hover::HoverParams;
 use pg_workspace::Workspace;
+use text_size::TextSize;
 use tokio::sync::RwLock;
 use tower_lsp::lsp_types::{
-    CodeAction, CodeActionOrCommand, CompletionItem, CompletionItemKind, CompletionList, Hover,
-    HoverContents, InlayHint, MarkedString, Position, Range,
+    CodeActionOrCommand, CompletionItem, CompletionItemKind, CompletionList, Hover, HoverContents, InlayHint, InlayHintKind, InlayHintLabel, MarkedString, Position, Range
 };
 
 use crate::{db_connection::DbConnection, utils::line_index_ext::LineIndexExt};
 
+#[derive(Clone)]
 pub struct Session {
-    db: RwLock<Option<DbConnection>>,
+    db: Arc<RwLock<Option<DbConnection>>>,
     ide: Arc<RwLock<Workspace>>,
 }
 
@@ -22,7 +24,7 @@ impl Session {
     pub fn new() -> Self {
         let ide = Arc::new(RwLock::new(Workspace::new()));
         Self {
-            db: RwLock::new(None),
+            db: Arc::new(RwLock::new(None)),
             ide,
         }
     }
@@ -64,16 +66,43 @@ impl Session {
     /// Runs the passed-in statement against the underlying database.
     pub async fn run_stmt(&self, stmt: String) -> anyhow::Result<u64> {
         let db = self.db.read().await;
-        let pool = db.map(|d| d.get_pool());
+        let pool = db.as_ref().map(|d| d.get_pool());
 
         let cmd = ExecuteStatementCommand::new(stmt);
 
-        cmd.run(pool).await
+        match cmd.run(pool).await {
+            Err(e) => Err(e),
+            Ok(res) => Ok(res.rows_affected()),
+        }
     }
 
     pub async fn on_file_closed(&self, path: PgLspPath) {
         let ide = self.ide.read().await;
         ide.remove_document(path);
+    }
+
+    pub fn get_diagnostics_sync(&self, path: PgLspPath) -> Vec<(Diagnostic, Range)> {
+        let ide = self.ide.blocking_read();
+
+        // make sure there are documents at the provided path before
+        // trying to collect diagnostics.
+        let doc = ide.documents.get(&path);
+        if doc.is_none() {
+            return vec![];
+        }
+
+        ide.diagnostics(&path)
+            .into_iter()
+            .map(|d| {
+                let range = doc
+                    .as_ref()
+                    .unwrap()
+                    .line_index
+                    .line_col_lsp_range(d.range)
+                    .unwrap();
+                (d, range)
+            })
+            .collect()
     }
 
     pub async fn get_diagnostics(&self, path: PgLspPath) -> Vec<(Diagnostic, Range)> {
@@ -105,7 +134,7 @@ impl Session {
         path: PgLspPath,
         version: i32,
         text: String,
-    ) -> HashSet<String> {
+    ) -> HashSet<PgLspPath> {
         {
             let ide = self.ide.read().await;
 
@@ -120,10 +149,10 @@ impl Session {
             );
         }
 
-        self.recompute_and_get_changed_files()
+        self.recompute_and_get_changed_files().await
     }
 
-    pub async fn recompute_and_get_changed_files(&self) -> HashSet<String> {
+    pub async fn recompute_and_get_changed_files(&self) -> HashSet<PgLspPath> {
         let ide = self.ide.read().await;
 
         let db = self.db.read().await;
@@ -133,7 +162,7 @@ impl Session {
 
         changed_files
             .into_iter()
-            .map(|f| f.document_url.to_string_lossy().to_string())
+            .map(|p| p.document_url)
             .collect()
     }
 
@@ -145,9 +174,6 @@ impl Session {
         let ide = self.ide.read().await;
         let doc = ide.documents.get(&path)?;
 
-        let db = self.db.read().await?;
-
-        let doc = doc.unwrap();
         let range = doc.line_index.offset_lsp_range(range).unwrap();
 
         // for now, we only provide `ExcecuteStatementCommand`s.
@@ -160,7 +186,7 @@ impl Session {
                     "Execute '{}'",
                     ExecuteStatementCommand::trim_statement(stmt.text.clone(), 50)
                 );
-                CodeActionOrCommand::Command(Command {
+                CodeActionOrCommand::Command(tower_lsp::lsp_types::Command {
                     title,
                     command: format!("pglsp.{}", cmd.id()),
                     arguments: Some(vec![serde_json::to_value(stmt.text.clone()).unwrap()]),
@@ -235,10 +261,10 @@ impl Session {
         let schema_cache = ide.schema_cache.read().expect("No Schema Cache");
 
         let completion_items = pg_completions::complete(&CompletionParams {
-            position: pos - range.start() - TextSize::from(1),
+            position: offset - range.start() - TextSize::from(1),
             text: stmt.text.as_str(),
             tree: ide.tree_sitter.tree(&stmt).as_ref().map(|x| x.as_ref()),
-            schema: &schema,
+            schema: &schema_cache,
         })
         .items
         .into_iter()
