+++ crates/pg_lsp/src/utils.rs
@@ -4,8 +4,8 @@ pub mod to_proto;
 
 use std::path::PathBuf;
 
-use lsp_types;
 use pg_base_db::PgLspPath;
+use tower_lsp::lsp_types;
 
 /// Convert a `lsp_types::Url` to a `PgLspPath`.
 pub(crate) fn file_path(url: &lsp_types::Url) -> PgLspPath {
